Pass an application instead of methods
In an effort to reduce the amount of boilerplate code that the consuming developer needs to write in order to integrate with Stormpath OmniAuth Strategy, an instance of the Application resource used to authenticate the user be available on the options object. Since both the Rails and SDK authenticate through an application - this should work regardless of whether or not the developer is using the 'stormpath-rails' or 'stormpath-sdk' gems.